projects/steganography-tool/requirements.txt
pillow
cryptography
tkinter